Moved more classes.
