+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gossip/FallenBehindManager.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.swirlds.platform.gossip.sync;
+package com.swirlds.platform.gossip;
 
 import com.swirlds.common.system.NodeId;
 import java.util.List;
