+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gossip/chatter/ChatterSyncProtocol.java
@@ -19,12 +19,12 @@ package com.swirlds.platform.gossip.chatter;
 import com.swirlds.common.system.NodeId;
 import com.swirlds.common.threading.pool.ParallelExecutionException;
 import com.swirlds.platform.Utilities;
+import com.swirlds.platform.gossip.FallenBehindManager;
+import com.swirlds.platform.gossip.SyncException;
 import com.swirlds.platform.gossip.chatter.communication.ChatterProtocol;
 import com.swirlds.platform.gossip.chatter.protocol.MessageProvider;
 import com.swirlds.platform.gossip.chatter.protocol.peer.CommunicationState;
-import com.swirlds.platform.gossip.sync.FallenBehindManager;
 import com.swirlds.platform.gossip.sync.ShadowGraphSynchronizer;
-import com.swirlds.platform.gossip.sync.SyncException;
 import com.swirlds.platform.network.Connection;
 import com.swirlds.platform.network.NetworkProtocolException;
 import com.swirlds.platform.network.protocol.Protocol;
