+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gossip/sync/SyncComms.java
@@ -20,6 +20,7 @@ import static com.swirlds.logging.LogMarker.SYNC_INFO;
 
 import com.swirlds.common.crypto.Hash;
 import com.swirlds.platform.event.GossipEvent;
+import com.swirlds.platform.gossip.SyncException;
 import com.swirlds.platform.internal.EventImpl;
 import com.swirlds.platform.metrics.SyncMetrics;
 import com.swirlds.platform.network.ByteConstants;
