+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gossip/sync/SyncInputStream.java
@@ -24,6 +24,7 @@ import com.swirlds.common.io.extendable.extensions.CountingStreamExtension;
 import com.swirlds.common.io.extendable.extensions.HashingStreamExtension;
 import com.swirlds.common.io.streams.SerializableDataInputStream;
 import com.swirlds.platform.event.GossipEvent;
+import com.swirlds.platform.gossip.SyncException;
 import com.swirlds.platform.network.ByteConstants;
 import java.io.BufferedInputStream;
 import java.io.IOException;
