+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gossip/sync/SyncManagerImpl.java
@@ -26,6 +26,7 @@ import com.swirlds.platform.Settings;
 import com.swirlds.platform.components.CriticalQuorum;
 import com.swirlds.platform.components.EventCreationRules;
 import com.swirlds.platform.event.EventIntakeTask;
+import com.swirlds.platform.gossip.FallenBehindManager;
 import com.swirlds.platform.network.RandomGraph;
 import java.util.LinkedList;
 import java.util.List;
