+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gossip/sync/SyncProtocolResponder.java
@@ -18,6 +18,8 @@ package com.swirlds.platform.gossip.sync;
 
 import com.swirlds.common.threading.locks.locked.MaybeLocked;
 import com.swirlds.common.threading.pool.ParallelExecutionException;
+import com.swirlds.platform.gossip.FallenBehindManager;
+import com.swirlds.platform.gossip.SyncException;
 import com.swirlds.platform.network.Connection;
 import com.swirlds.platform.network.NetworkProtocolException;
 import com.swirlds.platform.network.unidirectional.NetworkProtocolResponder;
