+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gossip/sync/protocol/SyncProtocol.java
@@ -25,9 +25,9 @@ import com.swirlds.common.threading.pool.ParallelExecutionException;
 import com.swirlds.common.time.Time;
 import com.swirlds.platform.Utilities;
 import com.swirlds.platform.components.CriticalQuorum;
-import com.swirlds.platform.gossip.sync.FallenBehindManager;
+import com.swirlds.platform.gossip.FallenBehindManager;
+import com.swirlds.platform.gossip.SyncException;
 import com.swirlds.platform.gossip.sync.ShadowGraphSynchronizer;
-import com.swirlds.platform.gossip.sync.SyncException;
 import com.swirlds.platform.metrics.SyncMetrics;
 import com.swirlds.platform.network.Connection;
 import com.swirlds.platform.network.NetworkProtocolException;
