+++ platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/sync/protocol/SyncProtocolTests.java
@@ -30,9 +30,9 @@ import com.swirlds.common.threading.SyncPermitProvider;
 import com.swirlds.common.threading.locks.locked.MaybeLocked;
 import com.swirlds.common.threading.pool.ParallelExecutionException;
 import com.swirlds.platform.components.CriticalQuorum;
-import com.swirlds.platform.gossip.sync.FallenBehindManager;
+import com.swirlds.platform.gossip.FallenBehindManager;
+import com.swirlds.platform.gossip.SyncException;
 import com.swirlds.platform.gossip.sync.ShadowGraphSynchronizer;
-import com.swirlds.platform.gossip.sync.SyncException;
 import com.swirlds.platform.gossip.sync.protocol.PeerAgnosticSyncChecks;
 import com.swirlds.platform.gossip.sync.protocol.SyncProtocol;
 import com.swirlds.platform.metrics.SyncMetrics;
