+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/test/java/com/swirlds/platform/test/sync/SyncExceptionTests.java
@@ -22,7 +22,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import com.swirlds.platform.gossip.sync.SyncException;
+import com.swirlds.platform.gossip.SyncException;
 import com.swirlds.platform.network.Connection;
 import org.junit.jupiter.api.Test;
