+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/test/java/com/swirlds/platform/test/sync/TestingSyncManager.java
@@ -17,7 +17,7 @@
 package com.swirlds.platform.test.sync;
 
 import com.swirlds.common.system.NodeId;
-import com.swirlds.platform.gossip.sync.FallenBehindManager;
+import com.swirlds.platform.gossip.FallenBehindManager;
 import com.swirlds.platform.gossip.sync.SyncManager;
 import com.swirlds.platform.gossip.sync.SyncResult;
 import java.util.List;
